# api/id/swagger.yml
openapi: 3.0.3
info:
  title: Authentication API
  description: |
    <b>Base URL:</b> <code>https://id.never-expires.com/</code><br>

    Error Handling
    API can run into two types of errors while processing a request: request acceptance errors and request processing errors:<br>
    If a request fails at the acceptance stage, API will return an HTTP response without a body. The response will include one of the following HTTP status codes:<br><br>
    
    • <b>401 Unauthorized:</b> This status is returned when the request lacks valid authentication credentials for the target resource<br>
    • <b>404 Not Found:</b> Requested resource could not be found on the server<br>
    • <b>405 Method Not Allowed:</b> Request method is known by the server but is not supported by the target resource<br>
    • <b>408 Request Timeout:</b> The API didn't complete the request within the expected time<br>
    • <b>500 Internal Server Error:</b> The server failed the request due to an internal error<br><br>
    
    <b>If an error occurs during the processing of a request the API will return the HTTP 422 status code in the header.</b><br> The body of the response will contain a JSON object with two fields: an internal API error code and a message.<br>
    The internal error code is a four-digit number. The first digit of the code indicates the category of the error, similar to HTTP status codes.<br>
    The list of errors:<br><br>
    
    <b>1xxx: HTTP Request Forming Errors</b><br>
...
    
...
servers:
...
paths:
...
  /user/email/confirm:
    get:
      summary: Email confirmation
      description: Confirm email if token from query is valid and not expired
      operationId: confirmEmail

      security:
        - confirmEmailToken: []
      responses:
        302:
          description: Request processed successfully and redirected to page with result status.
        422:
          description: |
            An API error occurred while processing the request. JSON contains an internal error status code describing the reason for the error and message.<br>
            Expected internal codes: 1003 MissingParameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        405:
          description: HTTP method is not allowed
        408:
          description: Timeout
        500:
          description: Unexpected server error
...
components:
  securitySchemes:
    authorizationHeader:
      type: apiKey
      in: header
      name: Authorization
      description: > 
        Can have format 'Bearer {token}' or just '{token}'
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access-jwt
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh-jwt
    sessionIDCookie:
      type: apiKey
      in: cookie
      name: session-id
    confirmEmailToken:
      type: apiKey
      in: query
      name: token
  schemas:
...
    ErrorMessage:
      description: Contains the internal status code and a message
      type: object
      properties:
        status_code:
          type: integer
        error:
          type: string